Basic uni test and exception handling

+++ src/main/java/com/journal/adib/Journal/Controllers/ResourceController.java
@@ -3,12 +3,15 @@ package com.journal.adib.Journal.Controllers;
 import com.journal.adib.Journal.Models.Resource;
 import com.journal.adib.Journal.Services.ResourceService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 
@@ -18,18 +21,22 @@ public class ResourceController {
     @Autowired
     ResourceService resourceService;
 
-//
-//    @GetMapping("/resources")
-//    public String find(Model model) {
-//        List<Resource> resources = resourceService.findAll();
-//        model.addAttribute("resources",resources);
-//        return "resource-list";
-//
-//    }
 
     @GetMapping("/resources")
-    public @ResponseBody List<Resource> find() {
-        return resourceService.findAll();
+    public String find(Model model) {
+        List<Resource> resources = resourceService.findAll();
+        model.addAttribute("resources",resources);
+        return "resource-list";
+
+    }
+
+    @GetMapping("/resources/{resourceId}")
+    public String find(@PathVariable(value="resourceId") Long resoureceId, Model model) {
+        Resource resource = resourceService.findById(resoureceId);
+        List<Resource> resources = new ArrayList<>();
+        resources.add(resource);
+        model.addAttribute("resources",resources);
+        return "resource-list";
 
     }
 
@@ -71,10 +78,22 @@ public class ResourceController {
         return resourceService.save(resource);
     }
 
-    @DeleteMapping("/resource/{resourceId}")
-    @ResponseBody
-    public void deleteById(@PathVariable(value="resourceId") Long resourceId){
-        resourceService.deleteById(resourceId);
+//    @PutMapping("/resources/{resourceId")
+//    public ResponseEntity<Resource> updateResource(@PathVariable Long resourceId){
+//        Optional<Resource> r = resourceService.findById(resourceId);
+//        if()
+//        if(Resource r = resourceService.findById(resourceId)){
+//
+//        }
+//    }
+
+    @DeleteMapping("/resources/{resourceId}")
+    public ResponseEntity<Resource> deleteResource(@PathVariable Long resourceId) {
+        Optional<Resource> resource = resourceService.findById(resourceId);
+        if(resource.isPresent()){
+            return resourceService.deleteById(resourceId);
+        }
+        return resourceService.deleteById(resourceId);
     }
